Adding tests for alternative IAsyncEnumerable approach.
This approach assumes that, much like setting up mocks on `Task.Result`,
one could set up mocks on a method that returns `IAsyncEnumerable`,
but by calling `ToEnumerable()` at the end to allow for `Returns()`
to have a valid value.

However, this doesn't work, as `... => ....ToEnumerable<string>()` is not supported by Moq today.